pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "onewheel-ha-bridge"
version = "0.1.0"
description = "Read-only VESC TCP to Home Assistant MQTT bridge for a custom Onewheel"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "OpenClaw" }]
dependencies = [
  "paho-mqtt>=2.1,<3",
]

[project.scripts]
onewheel-ha-bridge = "onewheel_ha_bridge.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
